Rename numberpicker to superwheelpicker
