+++ buildSrc/src/main/kotlin/ProjectConstants.kt
@@ -23,5 +23,6 @@ object ProjectConstants {
     const val HomeNotification = ":features:home:notification"
     const val RegisterSnsLogin = ":features:register:snslogin"
     const val RegisterOnboard = ":features:register:onboard"
-    const val XmlNumberPicker = ":xml:numberpicker"
+    const val XmlSuperWheelPicker = ":xml:superwheelpicker"
+    const val XmlPhotoEditor = ":xml:photoeditor"
 }
