+++ xml/superwheelpicker/src/main/AndroidManifest.xml
@@ -7,6 +7,6 @@
   ~ Please see: https://github.com/applemango-runnerbe/RunnerBe-Android/blob/main/LICENSE.
   -->
 
-<manifest package="team.applemango.runnerbe.xml.numberpicker">
+<manifest package="team.applemango.runnerbe.xml.superwheelpicker">
 
 </manifest>
